requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Database & ORM
mysqlclient==2.2.0
django-mysql==4.12.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
pyotp==2.9.0
cryptography==41.0.7

# Async Processing
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage & Media
boto3==1.34.0
django-storages==1.14.2
Pillow==10.1.0

# Search
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# API Documentation
drf-spectacular==0.26.5

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
sentry-sdk==1.38.0

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
uuid==1.30

# Development
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1